Fix bug for flush command
:Release Notes:
Fix the bug for When the "flush" command is executed.

:Detailed Notes:
The function DoCmdFlush() is trying to find context for
"PmLogCtl". There is any point to create context for "PmLogCtl".
So, when DoCmdFlush() is called, it'll create the context for
"pmlogctl" and will be record it's log to mainstream with new
PmLogInfo() api.

:Testing Performed:
Compiled locally and tested on locally and the device.
Verified in the log file, properly displayed in /var/log/messages.

:QA Notes:

:Issues Addressed:
[GF-11114] returnValue issue regarding of clearall API

Open-webOS-DCO-1.0-Signed-off-by: Hongjeon Ha <hj.ha@lge.com>

Change-Id: Ie4e410572c907a94992da8537358bd1ed10dcd5d
Reviewed-on: https://g2g.palm.com/2987
Reviewed-by: Build Verification
Reviewed-by: Hongjeon Ha <hj.ha@lge.com>
Tested-by: Hongjeon Ha <hj.ha@lge.com>
Reviewed-by: Keith Derrick <keith.derrick@lge.com>